+++ petit-felix.rb
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+require "petit-felix"
+
+### CLI interface for petitfelix
+
+felix = PetitFelix::Output.new(cl_args: ARGV)